HPCC-34453 Refactor the functions that access storage planes
