+++ dali/base/dautils.hpp
@@ -548,11 +548,7 @@ extern da_decl IPropertyTree * getDropZonePlane(const char * name);
 extern da_decl IPropertyTree * findPlane(const char *category, const char * path, const char * host, bool ipMatch, bool mustMatch);
 extern da_decl IPropertyTree * findDropZonePlane(const char * path, const char * host, bool ipMatch, bool mustMatch);
 extern da_decl bool validateDropZone(IPropertyTree *plane, const char *path, const char *host, bool ipMatch);
-extern da_decl unsigned getNumPlaneStripes(const char *clusterName);
-extern da_decl bool isHostInPlane(IPropertyTree *plane, const char *host, bool ipMatch);
-extern da_decl bool getPlaneHost(StringBuffer &host, IPropertyTree *plane, unsigned which);
-extern da_decl void getPlaneHosts(StringArray &hosts, IPropertyTree *plane);
-extern da_decl bool isPathInPlane(IPropertyTree *plane, const char *path);
+
 extern da_decl bool allowForeign();
 extern da_decl void setPageCacheTimeoutMilliSeconds(unsigned timeoutSeconds);
 extern da_decl void setMaxPageCacheItems(unsigned _maxPageCacheItems);
