+++ ecl/hthor/hthor.cpp
@@ -586,7 +586,7 @@ void CHThorDiskWriteActivity::resolve()
                 getDefaultStoragePlane(planeName);
             bool outputCompressionDefault = agent.queryWorkUnit()->getDebugValueBool("compressAllOutputs", isContainerized());
             outputPlaneCompressed = outputCompressionDefault;
-            Owned<IPropertyTree> plane = getStoragePlane(planeName);
+            Owned<const IPropertyTree> plane = getStoragePlaneConfig(planeName, false);
             if (plane)
                 outputPlaneCompressed = plane->getPropBool("@compressLogicalFiles", outputCompressionDefault);
         }
