+++ system/jlib/jfile.cpp
@@ -17,7 +17,6 @@
 #include "platform.h"
 
 #include <atomic>
-#include <array>
 #include <unordered_set>
 #include <unordered_map>
 
@@ -60,6 +59,7 @@
 #include "jmutex.hpp"
 #include "jfile.hpp"
 #include "jfile.ipp"
+#include "jplane.hpp"
 
 #include <limits.h>
 #include "jexcept.hpp"
@@ -7552,61 +7552,6 @@ IFileEventWatcher *createFileEventWatcher(FileWatchFunc callback)
 
 //---- Storage plane related functions ----------------------------------------------------
 
-IPropertyTree * getHostGroup(const char * name, bool required)
-{
-    if (!isEmptyString(name))
-    {
-        VStringBuffer xpath("storage/hostGroups[@name='%s']", name);
-        Owned<IPropertyTree> global = getGlobalConfig();
-        IPropertyTree * match = global->getPropTree(xpath);
-        if (match)
-            return match;
-    }
-    if (required)
-        throw makeStringExceptionV(-1, "No entry found for hostGroup: '%s'", name ? name : "<null>");
-    return nullptr;
-}
-
-IPropertyTree * getStoragePlane(const char * name)
-{
-    VStringBuffer xpath("storage/planes[@name='%s']", name);
-    Owned<IPropertyTree> global = getGlobalConfig();
-    return global->getPropTree(xpath);
-}
-
-IPropertyTree * getRemoteStorage(const char * name)
-{
-    VStringBuffer xpath("storage/remote[@name='%s']", name);
-    Owned<IPropertyTree> global = getGlobalConfig();
-    return global->getPropTree(xpath);
-}
-
-IPropertyTreeIterator * getRemoteStoragesIterator()
-{
-    return getGlobalConfigSP()->getElements("storage/remote");
-}
-
-IPropertyTreeIterator * getPlanesIterator(const char * category, const char *name)
-{
-    StringBuffer xpath("storage/planes");
-    if (!isEmptyString(category))
-        xpath.appendf("[@category='%s']", category);
-    if (!isEmptyString(name))
-        xpath.appendf("[@name='%s']", name);
-    return getGlobalConfigSP()->getElements(xpath);
-}
-
-IAPICopyClient * createApiCopyClient(IStorageApiInfo * source, IStorageApiInfo * target)
-{
-    ReadLockBlock block(containedFileHookLock);
-    ForEachItemIn(i, containedFileHooks)
-    {
-        IAPICopyClient * copyClient = containedFileHooks.item(i).getCopyApiClient(source, target);
-        if (copyClient)
-            return copyClient;
-    }
-    return nullptr;
-}
 
 
 // NB: This implementation is not thread-safe.
@@ -7709,98 +7654,27 @@ bool hasGenericFiletypeName(const char * name)
     return false;
 }
 
-///---------------------------------------------------------------------------------------------------------------------
+IAPICopyClient * createApiCopyClient(IStorageApiInfo * source, IStorageApiInfo * target)
+{
+    ReadLockBlock block(containedFileHookLock);
+    ForEachItemIn(i, containedFileHooks)
+    {
+        IAPICopyClient * copyClient = containedFileHooks.item(i).getCopyApiClient(source, target);
+        if (copyClient)
+            return copyClient;
+    }
+    return nullptr;
+}
+
+//---------------------------------------------------------------------------------------------------------------------
 
 // Cache/update plane attributes settings
 static unsigned jFileHookId = 0;
 
-
-// Declare the array with an anonymous struct
-enum PlaneAttrType { boolean, integer };
-struct PlaneAttributeInfo
-{
-    PlaneAttrType type;
-    size32_t scale;
-    bool isExpert;
-    const char *name;
-};
-static const std::array<PlaneAttributeInfo, PlaneAttributeCount> planeAttributeInfo = {{
-    { PlaneAttrType::integer, 1024, false, "blockedFileIOKB" },   // enum PlaneAttributeType::BlockedSequentialIO    {0}
-    { PlaneAttrType::integer, 1024, false, "blockedRandomIOKB" }, // enum PlaneAttributeType::blockedRandomIOKB      {1}
-    { PlaneAttrType::boolean, 0, true, "fileSyncWriteClose" },    // enum PlaneAttributeType::fileSyncWriteClose     {2}
-    { PlaneAttrType::boolean, 0, true, "concurrentWriteSupport" },// enum PlaneAttributeType::concurrentWriteSupport {3}
-    { PlaneAttrType::integer, 1, false, "writeSyncMarginMs" },    // enum PlaneAttributeType::WriteSyncMarginMs      {4}
-}};
-
-// {prefix, {key1: value1, key2: value2, ...}}
-typedef std::pair<std::string, std::array<unsigned __int64, PlaneAttributeCount>> PlaneAttributesMapElement;
-
-static std::unordered_map<std::string, PlaneAttributesMapElement> planeAttributesMap;
-static CriticalSection planeAttributeMapCrit;
-static constexpr unsigned __int64 unsetPlaneAttrValue = 0xFFFFFFFF00000000;
 MODULE_INIT(INIT_PRIORITY_STANDARD)
 {
     auto updateFunc = [&](const IPropertyTree *oldComponentConfiguration, const IPropertyTree *oldGlobalConfiguration)
     {
-        CriticalBlock b(planeAttributeMapCrit);
-        planeAttributesMap.clear();
-        Owned<IPropertyTreeIterator> planesIter = getPlanesIterator(nullptr, nullptr);
-        ForEach(*planesIter)
-        {
-            const IPropertyTree &plane = planesIter->query();
-            PlaneAttributesMapElement &element = planeAttributesMap[plane.queryProp("@name")];
-            const char * prefix = plane.queryProp("@prefix");
-            element.first = prefix ? prefix : ""; // If prefix is empty, avoid segfault by setting it to an empty string
-            auto &values = element.second;
-            for (unsigned propNum=0; propNum<PlaneAttributeType::PlaneAttributeCount; ++propNum)
-            {
-                const PlaneAttributeInfo &attrInfo = planeAttributeInfo[propNum];
-                std::string prop;
-                if (attrInfo.isExpert)
-                    prop += "expert/";
-                prop += "@" + std::string(attrInfo.name);
-                switch (attrInfo.type)
-                {
-                    case PlaneAttrType::integer:
-                    {
-                        unsigned __int64 value = plane.getPropInt64(prop.c_str(), unsetPlaneAttrValue);
-                        if (unsetPlaneAttrValue != value)
-                        {
-                            if (attrInfo.scale)
-                            {
-                                dbgassertex(PlaneAttrType::integer == attrInfo.type);
-                                value *= attrInfo.scale;
-                            }
-                        }
-                        values[propNum] = value;
-                        break;
-                    }
-                    case PlaneAttrType::boolean:
-                    {
-                        unsigned __int64 value;
-                        if (plane.hasProp(prop.c_str()))
-                            value = plane.getPropBool(prop.c_str()) ? 1 : 0;
-                        else if (FileSyncWriteClose == propNum) // temporary (if FileSyncWriteClose and unset, check legacy fileSyncMaxRetrySecs), purely for short term backward compatibility (see HPCC-32757)
-                        {
-                            unsigned __int64 v = plane.getPropInt64("expert/@fileSyncMaxRetrySecs", unsetPlaneAttrValue);
-                            // NB: fileSyncMaxRetrySecs==0 is treated as set/enabled
-                            if (unsetPlaneAttrValue != v)
-                                value = 1;
-                            else
-                                value = unsetPlaneAttrValue;
-                        }
-                        else
-                            value = unsetPlaneAttrValue;
-                        values[propNum] = value;
-                        break;
-                    }
-                    default:
-                        throwUnexpected();
-                }
-            }
-        }
-
-        // reset defaults
         expertEnableIFileFlagsMask = IFEnone;
         expertDisableIFileFlagsMask = IFEnone;
 
@@ -7817,7 +7691,7 @@ MODULE_INIT(INIT_PRIORITY_STANDARD)
         if (safeStatBehaviour >= SafeStatBehaviour::Count) // safeguard against bad config value
             safeStatBehaviour = defaultSafeStatBehaviour;
 
-        std::string propName = "expert/@" + std::string(planeAttributeInfo[FileSyncWriteClose].name);
+        std::string propName = "expert/@" + std::string(getPlaneAttributeString(FileSyncWriteClose));
         if (componentConfig->hasProp(propName.c_str()))
             defaultFileSyncWriteCloseEnabled = componentConfig->getPropBool(propName.c_str());
         else
@@ -7834,79 +7708,6 @@ MODULE_INIT(INIT_PRIORITY_STANDARD)
 MODULE_EXIT()
 {
     removeConfigUpdateHook(jFileHookId);
-}
-
-const char *getPlaneAttributeString(PlaneAttributeType attr)
-{
-    assertex(attr < PlaneAttributeCount);
-    return planeAttributeInfo[attr].name;
-}
-
-unsigned __int64 getPlaneAttributeValue(const char *planeName, PlaneAttributeType planeAttrType, unsigned __int64 defaultValue)
-{
-    if (!planeName)
-        return defaultValue;
-    assertex(planeAttrType < PlaneAttributeCount);
-    CriticalBlock b(planeAttributeMapCrit);
-    auto it = planeAttributesMap.find(planeName);
-    if (it != planeAttributesMap.end())
-    {
-        unsigned __int64 v = it->second.second[planeAttrType];
-        if (v != unsetPlaneAttrValue)
-            return v;
-    }
-    return defaultValue;
-}
-
-static PlaneAttributesMapElement *findPlaneElementFromPath(const char *filePath)
-{
-    for (auto &e: planeAttributesMap)
-    {
-        const char *prefix = e.second.first.c_str();
-        if (!isEmptyString(prefix)) // sanity check, std::string cannot be null, so check if empty
-        {
-            if (startsWith(filePath, prefix))
-                return &e.second;
-        }
-    }
-    return nullptr;
-}
-
-const char *findPlaneFromPath(const char *filePath, StringBuffer &result)
-{
-    CriticalBlock b(planeAttributeMapCrit);
-    PlaneAttributesMapElement *e = findPlaneElementFromPath(filePath);
-    if (!e)
-        return nullptr;
-
-    result.append(e->first.c_str());
-    return result;
-}
-
-bool findPlaneAttrFromPath(const char *filePath, PlaneAttributeType planeAttrType, unsigned __int64 defaultValue, unsigned __int64 &resultValue)
-{
-    CriticalBlock b(planeAttributeMapCrit);
-    PlaneAttributesMapElement *e = findPlaneElementFromPath(filePath);
-    if (e)
-    {
-        unsigned __int64 value = e->second[planeAttrType];
-        if (unsetPlaneAttrValue != value)
-            resultValue = value;
-        else
-            resultValue = defaultValue;
-        return true;
-    }
-    return false;
-}
-
-size32_t getBlockedFileIOSize(const char *planeName, size32_t defaultSize)
-{
-    return (size32_t)getPlaneAttributeValue(planeName, BlockedSequentialIO, defaultSize);
-}
-
-size32_t getBlockedRandomIOSize(const char *planeName, size32_t defaultSize)
-{
-    return (size32_t)getPlaneAttributeValue(planeName, BlockedRandomIO, defaultSize);
 }
 
 bool getFileSyncWriteCloseEnabled(const char *planeName)
